+++ weblate/trans/tests/test_component.py
@@ -261,24 +261,6 @@ class ComponentTest(RepoTestCase):
         project = self.create_link()
         self.verify_component(project, 3, 'cs', 4)
 
-    def test_extra_file(self):
-        """Extra commit file validation."""
-        project = self.create_component()
-        project.full_clean()
-
-        project.extra_commit_file = 'locale/list.txt'
-        project.full_clean()
-
-        project.extra_commit_file = 'locale/%(language)s.txt'
-        project.full_clean()
-
-        project.extra_commit_file = 'locale/%(bar)s.txt'
-        self.assertRaisesMessage(
-            ValidationError,
-            "Bad format string ('bar')",
-            project.full_clean
-        )
-
     def test_check_flags(self):
         """Check flags validation."""
         project = self.create_component()