+++ docs/changes.rst
@@ -11,6 +11,7 @@ Released on ? 2018.
 * The import_project management command has now slightly different parameters.
 * Added basic support for Windows RC files.
 * New addon to store contributor names in PO file headers.
+* The per component hook scripts are removed, use addons instead.
 
 weblate 2.20
 ------------